refactor: collect properties
