+++ four/src/iam/mod.rs
@@ -6,11 +6,12 @@ mod instance_profile;
 mod path;
 mod property;
 pub mod resource;
+mod role;
 
-pub use access_key::{AccessKeyId, AccessKeyStatus, SecretAccessKey};
-pub use arn::{GroupArn, InstanceProfileArn, ManagedPolicyArn, OIDCProviderArn};
+pub use access_key::{AccessKeyStatus, SecretAccessKey};
+pub use arn::{GroupArn, InstanceProfileArn, ManagedPolicyArn, OIDCProviderArn, RoleArn};
 pub use group::{GroupName, GroupNameError, Groups};
-pub use id::{GroupPolicyId, InstanceProfileId};
+pub use id::{AccessKeyId, GroupPolicyId, InstanceProfileId, PolicyId, RoleId, RolePolicyId};
 pub use instance_profile::{InstanceProfileName, InstanceProfileNameError};
 pub use path::{Path, PathError};
 pub use property::{
@@ -20,3 +21,4 @@ pub use property::{
     principal::{Principal, ServicePrincipal},
     statement::{ActionOr, PrincipalOr, Statement, StatementBuilder1, StatementBuilder2},
 };
+pub use role::RoleName;
