+++ four/src/iam/resource/role.rs
@@ -1,16 +1,18 @@
 use crate::{
     core::{
-        convert::{WillBe, WillMappable},
+        convert::WillBe,
         function::{HaveAtt, RefInner, Referenced},
         service::IAM,
         Arn, LogicalId,
     },
-    iam::property::{
-        action, policy_document::PolicyDocument, principal::Principal, statement::Statement,
+    iam::{
+        property::{
+            action, policy_document::PolicyDocument, principal::Principal, statement::Statement,
+        },
+        RoleArn, RoleId, RoleName,
     },
 };
 use four_derive::ManagedResource;
-use serde::Serialize;
 
 #[derive(ManagedResource, Clone)]
 #[resource_type = "AWS::IAM::Role"]
@@ -40,29 +42,6 @@ impl Referenced for Role {
         RefInner::Id(self.logical_id.clone())
     }
 }
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RoleName(String);
-
-impl RoleName {
-    pub fn new(name: String) -> Self {
-        Self(name)
-    }
-}
-
-impl WillMappable<String> for RoleName {}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RoleArn(Arn<IAM>);
-
-impl From<Arn<IAM>> for RoleArn {
-    fn from(value: Arn<IAM>) -> RoleArn {
-        RoleArn(value)
-    }
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct RoleId(String);
 
 impl HaveAtt<RoleArn> for Role {
     const KEY: &'static str = "Arn";
